chore(api): move shared read-models unit tests in shared
