+++ api/tests/shared/unit/domain/read-models/participant-results/AssessmentResult_test.js
@@ -4,12 +4,12 @@ import sinon from 'sinon';
 import {
   CampaignParticipationStatuses,
   CampaignTypes,
-} from '../../../../../src/prescription/shared/domain/constants.js';
-import { constants } from '../../../../../src/shared/domain/constants.js';
-import { KnowledgeElement } from '../../../../../src/shared/domain/models/KnowledgeElement.js';
-import { AssessmentResult } from '../../../../../src/shared/domain/read-models/participant-results/AssessmentResult.js';
-import { expect } from '../../../../test-helper.js';
-import { domainBuilder } from '../../../../tooling/domain-builder/domain-builder.js';
+} from '../../../../../../src/prescription/shared/domain/constants.js';
+import { constants } from '../../../../../../src/shared/domain/constants.js';
+import { KnowledgeElement } from '../../../../../../src/shared/domain/models/KnowledgeElement.js';
+import { AssessmentResult } from '../../../../../../src/shared/domain/read-models/participant-results/AssessmentResult.js';
+import { expect } from '../../../../../test-helper.js';
+import { domainBuilder } from '../../../../../tooling/domain-builder/domain-builder.js';
 
 describe('Unit | Domain | Read-Models | ParticipantResult | AssessmentResult', function () {
   describe('#constructor', function () {
