+++ api/tests/shared/unit/domain/read-models/participant-results/CompetenceResult_test.js
@@ -1,7 +1,7 @@
-import { KnowledgeElement } from '../../../../../src/shared/domain/models/KnowledgeElement.js';
-import { CompetenceResult } from '../../../../../src/shared/domain/read-models/participant-results/CompetenceResult.js';
-import { expect } from '../../../../test-helper.js';
-import { domainBuilder } from '../../../../tooling/domain-builder/domain-builder.js';
+import { KnowledgeElement } from '../../../../../../src/shared/domain/models/KnowledgeElement.js';
+import { CompetenceResult } from '../../../../../../src/shared/domain/read-models/participant-results/CompetenceResult.js';
+import { expect } from '../../../../../test-helper.js';
+import { domainBuilder } from '../../../../../tooling/domain-builder/domain-builder.js';
 
 describe('Unit | Domain | Read-Models | ParticipantResult | CompetenceResult', function () {
   it('computes the result for the given competence', function () {
